backend/requirements.txt

fastapi
uvicorn[standard]
python-dotenv
groq
pytesseract
Pillow
PyPDF2
python-pptx
